% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{GLMsim_cell_info}
\alias{GLMsim_cell_info}
\title{Collecting cell information from the original experimental data}
\usage{
GLMsim_cell_info(real_count, biology_vec = NULL, batch_vec = NULL)
}
\arguments{
\item{real_count}{real data count matrix. Rows: genes. Columns: cells.}

\item{biology_vec}{a factor vector of biology groups.}

\item{batch_vec}{a factor vector of batches.}
}
\value{
A data frame to reorganize the information for biology and unwanted variations.
}
\description{
This function collects information from the original experimental data. The library size, biological and batch information are extracted from the original experimental data, which will be prepared for fitting each gene into a generalized linear model for the next step.
}
