% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{GLMsim_fit}
\alias{GLMsim_fit}
\title{Fitting each gene count into a generalized linear model.}
\usage{
GLMsim_fit(real_count, cell_info, parallel = TRUE, ncores = "all")
}
\arguments{
\item{real_count}{real data count matrix. Rows: genes. Columns: cells.}

\item{cell_info}{a data frame of cell information that outputs from data_prepare function.}

\item{parallel}{whether to fit all genes into GLM in parallel. Default: TRUE.}

\item{ncores}{number of cores that can be used for parallel computing. Default: "all", which means using all cores. If that does not work, try 6 or 8.}
}
\value{
A matrix of fitted coefficients from the real experimental data. In this matrix, each row represents a gene, and each column corresponds to a coefficient.
}
\description{
This function fits UMI count for each gene into a generalized linear model, which aims to obtain the estimated coefficients for biological effect and unwanted variation.
}
