% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_outlier.R
\name{GLMsim_sim_outlier}
\alias{GLMsim_sim_outlier}
\title{Simulating counts with outlier genes.}
\usage{
GLMsim_sim_outlier(
  real_count,
  cell_info,
  coe.mat,
  sim.glm.count,
  fit.method = "trend",
  nmad = 30,
  parallel = FALSE,
  trunc_q = c(0.05, 0.95)
)
}
\arguments{
\item{real_count}{real data count matrix. Rows: genes. Columns: cells.}

\item{cell_info}{output data frame from data_prepare function.}

\item{coe.mat}{output of glm fitted coefficients from GLMsim_fit function.}

\item{sim.glm.count}{the initial simulated count matrix from GLMsim_simulation function.}

\item{fit.method}{the method used to fit outlier genes. Users should choose one of the methods from “robust_glm” or “clipping” or “trend”. Default: “trend” method.}

\item{nmad}{number of median absolute deviation (MAD) to determine outlier genes. Default: 30.}

\item{parallel}{whether to fit outlier genes into GLM in parallel or not. Default: FALSE.}

\item{trunc_q}{a vector consisting of 2 elements to indicate the quantile range for the cut-off of the Winsorizing method. The default is c(0.05, 0.95).}
}
\value{
A simulated single cell count matrix after correcting outlier genes, with rows corresponding to genes and columns corresponding to cells. If no outlier genes exist, the output simulated count will be same as that from the GLMsim_simulation function.
}
\description{
To mitigate the effects of extreme counts from outlier genes, this function will check if outlier genes exist at first. If outlier genes exist in the initial simulated data, this function allows users to choose one of three methods: robust negative binomial GLM, Winsorizing and trended coefficient to refit those outlier genes. Then simulate counts for those outlier genes using refitted coefficients.
}
