% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_count.R
\name{GLMsim_simulation}
\alias{GLMsim_simulation}
\title{Simulating single cell counts using estimated coefficients.}
\usage{
GLMsim_simulation(coe.mat, cell_info, seed = 2022)
}
\arguments{
\item{coe.mat}{output of glm fitted coefficients from GLMsim_fit function.}

\item{cell_info}{output data frame from data_prepare function.}

\item{seed}{used for set.seed() to generate random numbers. Default: 2022.}
}
\value{
An initial simulated single cell count matrix, with rows corresponding to genes and columns corresponding to cells.
}
\description{
This function performs an initial simulation of single cell counts using fitted coefficients from the previous step.
}
